Refactored String Literals to Constants
